This update make map rendering a bit more logical (and uses a dirty trick)
